+++ src/android/Wechat.java
@@ -39,6 +39,9 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.ByteArrayInputStream;
+import org.apache.cordova.CordovaPreferences;
 
 public class Wechat extends CordovaPlugin {
 
@@ -96,13 +99,14 @@ public class Wechat extends CordovaPlugin {
     protected static CallbackContext currentCallbackContext;
     protected static IWXAPI wxAPI;
     protected static String appId;
+    protected static CordovaPreferences wx_preferences;
 
     @Override
     protected void pluginInitialize() {
 
         super.pluginInitialize();
 
-        String id = getAppId();
+        String id = getAppId(preferences);
 
         // save app id
         saveAppId(cordova.getActivity(), id);
@@ -115,9 +119,11 @@ public class Wechat extends CordovaPlugin {
 
     protected void initWXAPI() {
         IWXAPI api = getWxAPI(cordova.getActivity());
-
+        if(wx_preferences == null) {
+            wx_preferences = preferences;
+        }
         if (api != null) {
-            api.registerApp(getAppId());
+            api.registerApp(getAppId(preferences));
         }
     }
 
@@ -278,7 +284,7 @@ public class Wechat extends CordovaPlugin {
 
         try {
             final String appid = params.getString("appid");
-            final String savedAppid = getAppId(cordova.getActivity());
+            final String savedAppid = getSavedAppId(cordova.getActivity());
             if (!savedAppid.equals(appid)) {
                 this.saveAppId(cordova.getActivity(), appid);
             }
@@ -330,7 +336,7 @@ public class Wechat extends CordovaPlugin {
                ChooseCardFromWXCardPackage.Req req = new ChooseCardFromWXCardPackage.Req();
 
                try {
-                   req.appId = getAppId();
+                   req.appId = getAppId(preferences);
                    req.cardType = "INVOICE";
                    req.signType = params.getString("signType");
                    req.cardSign = params.getString("cardSign");
@@ -514,6 +520,12 @@ public class Wechat extends CordovaPlugin {
                 Bitmap scaled = Bitmap.createScaledBitmap(bmp, width, height, true);
                 bmp.recycle();
 
+                int length = scaled.getRowBytes() * scaled.getHeight();
+
+                if(length > (maxSize/10)*1024) {
+                    scaled = compressImage(scaled,(maxSize/10));
+                }
+
                 bmp = scaled;
             }
 
@@ -530,11 +542,28 @@ public class Wechat extends CordovaPlugin {
         return bmp;
     }
 
+
+    /**
+     * compress bitmap by quility
+     */
+    protected  Bitmap compressImage(Bitmap image,Integer maxSize) {
+
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        image.compress(Bitmap.CompressFormat.JPEG, 100, baos);
+        int options = 90;
+
+        while (baos.toByteArray().length / 1024 > maxSize) { 
+            baos.reset(); 
+            image.compress(Bitmap.CompressFormat.JPEG, options, baos);
+            options -= 10;
+        }
+        ByteArrayInputStream isBm = new ByteArrayInputStream(baos.toByteArray());
+        Bitmap bitmap = BitmapFactory.decodeStream(isBm, null, null);
+        return bitmap;
+    }
+
     /**
      * Get input stream from a url
-     *
-     * @param url
-     * @return
      */
     protected InputStream getFileInputStream(String url) {
         try {
@@ -595,9 +624,13 @@ public class Wechat extends CordovaPlugin {
         return null;
     }
 
-    public static String getAppId() {
+    public static String getAppId(CordovaPreferences f_preferences) {
         if (appId == null) {
-            appId = preferences.getString(WXAPPID_PROPERTY_KEY, "");
+            if(f_preferences != null) {
+                appId = f_preferences.getString(WXAPPID_PROPERTY_KEY, "");
+            }else if(wx_preferences != null){
+                appId = wx_preferences.getString(WXAPPID_PROPERTY_KEY, "");
+            }
         }
 
         return appId;
@@ -619,7 +652,7 @@ public class Wechat extends CordovaPlugin {
      * @param id
      */
     public static void saveAppId(Context ctx, String id) {
-        if (id.isEmpty()) {
+        if (id!=null && id.isEmpty()) {
             return ;
         }